requirements.txt
lmdb
numpy
tqdm
pillow
opencv-python
torch-fidelity
nltk
einops
easydict